requirements.txt
Django==1.8.3
gunicorn==19.3.0
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
wheel==0.24.0
python-documentcloud==1.0.2
